rename and move some files to another packages. remove some un used files and code.
