+++ app/src/main/java/com/amsterdam/cutetudee/data/service/TaskServiceImpl.kt
@@ -5,8 +5,8 @@ import com.amsterdam.cutetudee.data.mapper.toLong
 import com.amsterdam.cutetudee.data.mapper.toTask
 import com.amsterdam.cutetudee.data.mapper.toTaskEntity
 import com.amsterdam.cutetudee.data.mapper.toTaskListFlow
-import com.amsterdam.cutetudee.domain.model.Task
-import com.amsterdam.cutetudee.domain.model.TaskStatistics
+import com.amsterdam.cutetudee.domain.entity.Task
+import com.amsterdam.cutetudee.domain.entity.TaskStatistics
 import com.amsterdam.cutetudee.domain.service.TaskService
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
