+++ app/src/main/java/com/amsterdam/cutetudee/di/AppModule.kt
@@ -4,5 +4,4 @@ val appModule = listOf(
     dataModule,
     serviceModule,
     viewModelModule,
-    utilsModule
 )
