--- app/src/main/java/com/amsterdam/cutetudee/di/UtilsModule.kt
@@ -1,7 +1,0 @@
-package com.amsterdam.cutetudee.di
-
-import org.koin.dsl.module
-
-val utilsModule = module {
-
-}
