+++ app/src/main/java/com/amsterdam/cutetudee/di/ViewModelModule.kt
@@ -1,6 +1,6 @@
 package com.amsterdam.cutetudee.di
 
-import com.amsterdam.cutetudee.MainViewModel
+import com.amsterdam.cutetudee.presentation.AppViewModel
 import com.amsterdam.cutetudee.presentation.screens.category.CategoryViewModel
 import com.amsterdam.cutetudee.presentation.screens.categoryDetails.CategoryDetailsViewModel
 import com.amsterdam.cutetudee.presentation.screens.home.HomeViewModel
@@ -8,7 +8,6 @@ import com.amsterdam.cutetudee.presentation.screens.onBoarding.OnBoardingViewMod
 import com.amsterdam.cutetudee.presentation.screens.splash.SplashViewModel
 import com.amsterdam.cutetudee.presentation.screens.tasks.AddEditTaskViewModel
 import com.amsterdam.cutetudee.presentation.screens.tasks.TasksViewModel
-import org.koin.android.ext.koin.androidContext
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
@@ -18,7 +17,7 @@ val viewModelModule = module {
     viewModelOf(::TasksViewModel)
     viewModelOf(::CategoryViewModel)
     viewModelOf(::HomeViewModel)
-    viewModelOf(::MainViewModel)
+    viewModelOf(::AppViewModel)
     viewModelOf(::CategoryDetailsViewModel)
     viewModelOf(::AddEditTaskViewModel)
 }
