+++ app/src/main/java/com/amsterdam/cutetudee/domain/entity/Task.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.cutetudee.domain.model
+package com.amsterdam.cutetudee.domain.entity
 
 import kotlinx.datetime.LocalDate
 import kotlin.uuid.ExperimentalUuidApi
