+++ app/src/main/java/com/amsterdam/cutetudee/presentation/component/chip/priority/PriorityChip.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import com.amsterdam.cutetudee.domain.model.Task
+import com.amsterdam.cutetudee.domain.entity.Task
 import com.amsterdam.cutetudee.presentation.component.chip.BasicChip
 import com.amsterdam.cutetudee.presentation.theme.AppTheme
