+++ app/src/main/java/com/amsterdam/cutetudee/presentation/component/chip/tast_status/TaskStatusChip.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.amsterdam.cutetudee.domain.model.Task
+import com.amsterdam.cutetudee.domain.entity.Task
 import com.amsterdam.cutetudee.presentation.component.chip.BasicChip
 import com.amsterdam.cutetudee.presentation.theme.AppTheme
