+++ app/src/main/java/com/amsterdam/cutetudee/presentation/screens/category/mappers/CategoryMapper.kt
@@ -1,11 +1,8 @@
 package com.amsterdam.cutetudee.presentation.screens.category.mappers
 
-import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.graphics.painter.BitmapPainter
 import androidx.core.net.toUri
-import com.amsterdam.cutetudee.domain.model.Category
+import com.amsterdam.cutetudee.domain.entity.Category
 import com.amsterdam.cutetudee.presentation.screens.category.CategoryUiState
-import com.amsterdam.cutetudee.presentation.utils.toBitmap
 import kotlin.uuid.ExperimentalUuidApi
 
 @OptIn(ExperimentalUuidApi::class)
