+++ app/src/main/java/com/amsterdam/cutetudee/presentation/screens/categoryDetails/CategoryDetailsInteractionListener.kt
@@ -1,7 +1,7 @@
 package com.amsterdam.cutetudee.presentation.screens.categoryDetails
 
 import android.net.Uri
-import com.amsterdam.cutetudee.domain.model.Task
+import com.amsterdam.cutetudee.domain.entity.Task
 
 interface CategoryDetailsInteractionListener {
     fun onTaskStatusChanged(taskStatus: Task.Status)
