+++ app/src/main/java/com/amsterdam/cutetudee/presentation/screens/home/HomeUiState.kt
@@ -4,7 +4,7 @@ import android.net.Uri
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import com.amsterdam.cutetudee.R
-import com.amsterdam.cutetudee.domain.model.Task
+import com.amsterdam.cutetudee.domain.entity.Task
 import com.amsterdam.cutetudee.presentation.component.chip.priority.PriorityUi
 import com.amsterdam.cutetudee.presentation.model.TaskUi
 import kotlinx.datetime.Clock
