--- app/src/main/java/com/amsterdam/cutetudee/presentation/utils/BitmapMapper.kt
@@ -1,19 +1,0 @@
-package com.amsterdam.cutetudee.presentation.utils
-
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
-import android.util.Base64
-import java.io.ByteArrayOutputStream
-
-fun Bitmap.toBase46eString(): String {
-    val outputStream = ByteArrayOutputStream()
-    this.compress(Bitmap.CompressFormat.PNG, 100, outputStream)
-    val byteArray = outputStream.toByteArray()
-    return Base64.encodeToString(byteArray, Base64.DEFAULT)
-}
-
-
-fun String.toBitmap(): Bitmap {
-    val decodedBytes = Base64.decode(this, Base64.DEFAULT)
-    return BitmapFactory.decodeByteArray(decodedBytes, 0, decodedBytes.size)
-}
