--- app/src/main/java/com/amsterdam/cutetudee/presentation/utils/NoRippleInteractionSource.kt
@@ -1,15 +1,0 @@
-package com.amsterdam.cutetudee.presentation.utils
-
-import androidx.compose.foundation.interaction.Interaction
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.emptyFlow
-
-object NoRippleInteractionSource : MutableInteractionSource {
-
-    override val interactions: Flow<Interaction> = emptyFlow()
-
-    override suspend fun emit(interaction: Interaction) {}
-
-    override fun tryEmit(interaction: Interaction) = true
-}
