--- app/src/main/java/com/amsterdam/cutetudee/presentation/utils/PresentationConstants.kt
@@ -1,7 +1,0 @@
-package com.amsterdam.cutetudee.presentation.utils
-
-object PresentationConstants {
-    object AnimationConstants {
-         const val NAVIGATION_ANIMATION_DURATION = 600
-    }
-}
